+++ chances 2.0/Assets/Scripts/Gluttony/TimerGame.cs
@@ -12,7 +12,7 @@ public class TimerGame : MonoBehaviour
     [SerializeField] private HealthSystem healthSystem;
     [SerializeField] private AttackGluttony attackGluttony;
     [SerializeField] private StartBlinkingAnim blink;
-    [SerializeField] private float health;
+    [SerializeField] private int health;
 
     public float spawnTimer = 10f;
 
@@ -40,7 +40,7 @@ public class TimerGame : MonoBehaviour
 
                 if (health > 0)
                 {
-                    float rnd = Random.Range(20, 30);
+                    int rnd = Random.Range(20, 30);
                     attackGluttony.ReturnAll();
                     healthSystem.TakeDamage(rnd);
                     blink.StartBlinking(0);
@@ -48,7 +48,7 @@ public class TimerGame : MonoBehaviour
                 else
                 {
                     // dmg player
-                    float rnd = Random.Range(10, 20);
+                    int rnd = Random.Range(10, 20);
                     healthSystemPlayer.TakeDamage(rnd);
                     attackGluttony.ReturnAll();
                     blink.StartBlinking(1);
